fix(vim): remove submodules
Removes vim-buftabline because it was irrelevant. Caused confusion and
was not actually useful. Statusline shows up from time to time anyways,
so I don't see a point to use buftabline.
Removes multiple cursors support. It drove me nuts sometimes because of
the lag. `Far` replaces it anyways. In a much cleaner way.